backend stuff

+++ migrations/2025-05-18-031123_create_comments/up.sql
@@ -2,8 +2,8 @@ CREATE TABLE comments (
   id INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT UNIQUE,
   torrent_id INTEGER NOT NULL,
   submitter VARCHAR NOT NULL,
-  date_created INTEGER NOT NULL,
-  date_edited INTEGER,
+  date_created BIGINT NOT NULL,
+  date_edited BIGINT,
   text VARCHAR NOT NULL,
   CONSTRAINT fk_torrents
     FOREIGN KEY (torrent_id)
